server/requirements.txt
asgiref==3.7.2
Django==5.0.9
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
python-decouple==3.8

pip~=24.2
wheel~=0.44.0
setuptools~=68.2.0